+++ src/Bridge/ElevenLabs/CHANGELOG.md
@@ -1,6 +1,13 @@
 CHANGELOG
 =========
 
+0.5
+---
+
+ * [BC BREAK] The `hostUrl` parameter for `ElevenLabsClient` has been removed
+ * [BC BREAK] The `host` parameter for `ElevenLabsApiCatalog` has been removed
+ * [BC BREAK] The `hostUrl` parameter for `PlatformFactory::create()` has been renamed to `endpoint`
+
 0.3
 ---